Extracted IoSubViewModel
