+++ PixiEditor/ViewModels/SubViewModels/Main/IoViewModel.cs
@@ -1,171 +1,103 @@
-﻿using Microsoft.Win32;
-using PixiEditor.Helpers;
-using PixiEditor.Models.DataHolders;
-using PixiEditor.Models.Dialogs;
-using PixiEditor.Models.Enums;
-using PixiEditor.Models.IO;
+﻿using PixiEditor.Helpers;
+using PixiEditor.Models.Controllers.Shortcuts;
+using PixiEditor.Models.Position;
 using System;
-using System.IO;
-using System.Linq;
 using System.Windows;
-using System.Windows.Media.Imaging;
+using System.Windows.Input;
 
 namespace PixiEditor.ViewModels.SubViewModels.Main
 {
     public class IoViewModel : SubViewModel<ViewModelMain>
     {
-        public RelayCommand OpenNewFilePopupCommand { get; set; }
-        public RelayCommand SaveDocumentCommand { get; set; }
-        public RelayCommand OpenFileCommand { get; set; }
-        public RelayCommand ExportFileCommand { get; set; } //Command that is used to save file
+        public RelayCommand MouseMoveCommand { get; set; }
+        public RelayCommand MouseDownCommand { get; set; }
+        public RelayCommand KeyDownCommand { get; set; }
+        public RelayCommand KeyUpCommand { get; set; }
+
+        private bool _restoreToolOnKeyUp = false;
 
         public IoViewModel(ViewModelMain owner) : base(owner)
         {
-            OpenNewFilePopupCommand = new RelayCommand(OpenNewFilePopup);
-            SaveDocumentCommand = new RelayCommand(SaveDocument, Owner.DocumentIsNotNull);
-            OpenFileCommand = new RelayCommand(Open);
-            ExportFileCommand = new RelayCommand(ExportFile, CanSave);
-            Owner.OnStartupEvent += Owner_OnStartupEvent;
+            MouseMoveCommand = new RelayCommand(MouseMove);
+            MouseDownCommand = new RelayCommand(MouseDown);
+            KeyDownCommand = new RelayCommand(KeyDown);
+            KeyUpCommand = new RelayCommand(KeyUp);
         }
 
-        private void Owner_OnStartupEvent(object sender, System.EventArgs e)
+        public void MouseHook_OnMouseUp(object sender, Point p, MouseButton button)
         {
-            var lastArg = Environment.GetCommandLineArgs().Last();
-            if (Importer.IsSupportedFile(lastArg) && File.Exists(lastArg))
+            GlobalMouseHook.OnMouseUp -= MouseHook_OnMouseUp;
+            if (button == MouseButton.Left)
             {
-                Open(lastArg);
+                Owner.BitmapManager.MouseController.StopRecordingMouseMovementChanges();
             }
-            else
-            {
-                OpenNewFilePopup(null);
-            }
+            Owner.BitmapManager.MouseController.MouseUp(new MouseEventArgs(Mouse.PrimaryDevice,
+                (int)DateTimeOffset.UtcNow.ToUnixTimeSeconds()));
         }
 
-        /// <summary>
-        ///     Generates new Layer and sets it as active one
-        /// </summary>
-        /// <param name="parameter"></param>
-        public void OpenNewFilePopup(object parameter)
+        private void MouseDown(object parameter)
         {
-            NewFileDialog newFile = new NewFileDialog();
-            if (newFile.ShowDialog())
+            if (Owner.BitmapManager.ActiveDocument.Layers.Count == 0) return;
+            if (Mouse.LeftButton == MouseButtonState.Pressed)
             {
-                NewDocument(newFile.Width, newFile.Height);
-            }
-        }
-
-        public void NewDocument(int width, int height, bool addBaseLayer = true)
-        {
-            Owner.BitmapManager.ActiveDocument = new Document(width, height);
-            if (addBaseLayer)
-            {
-                Owner.BitmapManager.AddNewLayer("Base Layer");
-            }
-            Owner.ResetProgramStateValues();
-        }
-
-
-        /// <summary>
-        ///     Opens file from path.
-        /// </summary>
-        /// <param name="path"></param>
-        public void OpenFile(string path)
-        {
-            ImportFileDialog dialog = new ImportFileDialog();
-
-            if (path != null && File.Exists(path))
-                dialog.FilePath = path;
-
-            if (dialog.ShowDialog())
-            {
-                NewDocument(dialog.FileWidth, dialog.FileHeight, false);
-                Owner.BitmapManager.AddNewLayer("Image", Importer.ImportImage(dialog.FilePath, dialog.FileWidth, dialog.FileHeight));
-            }
-        }
-
-        private void Open(string path)
-        {
-            if (Owner.UnsavedDocumentModified)
-            {
-                var result = ConfirmationDialog.Show(ViewModelMain.ConfirmationDialogMessage);
-                if (result == ConfirmationType.Yes)
+                if (!Owner.BitmapManager.MouseController.IsRecordingChanges)
                 {
-                    SaveDocument(null);
-                }
-                else if (result == ConfirmationType.Canceled)
-                {
-                    return;
+                    bool clickedOnCanvas = Owner.MouseXOnCanvas >= 0 &&
+                        Owner.MouseXOnCanvas <= Owner.BitmapManager.ActiveDocument.Width &&
+                        Owner.MouseYOnCanvas >= 0 && 
+                        Owner.MouseYOnCanvas <= Owner.BitmapManager.ActiveDocument.Height;
+                    Owner.BitmapManager.MouseController.StartRecordingMouseMovementChanges(clickedOnCanvas);
+                    Owner.BitmapManager.MouseController.RecordMouseMovementChange(MousePositionConverter.CurrentCoordinates);
                 }
             }
+            Owner.BitmapManager.MouseController.MouseDown(new MouseEventArgs(Mouse.PrimaryDevice,
+                (int)DateTimeOffset.UtcNow.ToUnixTimeSeconds()));
 
-            Owner.ResetProgramStateValues();
-            if (path.EndsWith(".pixi"))
-                OpenDocument(path);
-            else
-                OpenFile(path);
-        }
-
-        private void Open(object property)
-        {
-            OpenFileDialog dialog = new OpenFileDialog
-            {
-                Filter = "All Files|*.*|PixiEditor Files | *.pixi|PNG Files|*.png",
-                DefaultExt = "pixi"
-            };
-            if ((bool)dialog.ShowDialog())
-            {
-                if (Importer.IsSupportedFile(dialog.FileName))
-                    Open(dialog.FileName);
-                Owner.RecenterZoombox = !Owner.RecenterZoombox;
-            }
-        }
-
-        private void OpenDocument(string path)
-        {
-            Owner.BitmapManager.ActiveDocument = Importer.ImportDocument(path);
-            Exporter.SaveDocumentPath = path;
-            Owner.UnsavedDocumentModified = false;
-        }
-
-        public void SaveDocument(bool asNew)
-        {
-            SaveDocument(parameter: asNew ? "asnew" : null);
-        }
-
-        private void SaveDocument(object parameter)
-        {
-            bool paramIsAsNew = parameter != null && parameter.ToString()?.ToLower() == "asnew";
-            if (paramIsAsNew || Exporter.SaveDocumentPath == null)
-            {
-                var saved = Exporter.SaveAsEditableFileWithDialog(Owner.BitmapManager.ActiveDocument, !paramIsAsNew);
-                Owner.UnsavedDocumentModified = Owner.UnsavedDocumentModified && !saved;
-            }
-            else
-            {
-                Exporter.SaveAsEditableFile(Owner.BitmapManager.ActiveDocument, Exporter.SaveDocumentPath);
-                Owner.UnsavedDocumentModified = false;
-            }
+            // Mouse down is guaranteed to only be raised from within this application, so by subscribing here we
+            // only listen for mouse up events that occurred as a result of a mouse down within this application.
+            // This seems better than maintaining a global listener indefinitely.
+            GlobalMouseHook.OnMouseUp += MouseHook_OnMouseUp;
         }
 
         /// <summary>
-        ///     Generates export dialog or saves directly if save data is known.
+        ///     Method connected with command, it executes tool "activity"
         /// </summary>
         /// <param name="parameter"></param>
-        private void ExportFile(object parameter)
+        private void MouseMove(object parameter)
         {
-            WriteableBitmap bitmap = Owner.BitmapManager.GetCombinedLayersBitmap();
-            Exporter.Export(bitmap, new Size(bitmap.PixelWidth, bitmap.PixelHeight));
+            Coordinates cords = new Coordinates((int)Owner.MouseXOnCanvas, (int)Owner.MouseYOnCanvas);
+            MousePositionConverter.CurrentCoordinates = cords;
+
+
+            if (Owner.BitmapManager.MouseController.IsRecordingChanges && Mouse.LeftButton == MouseButtonState.Pressed)
+            {
+                Owner.BitmapManager.MouseController.RecordMouseMovementChange(cords);
+            }
+            Owner.BitmapManager.MouseController.MouseMoved(cords);
         }
 
-        /// <summary>
-        ///     Returns true if file save is possible.
-        /// </summary>
-        /// <param name="property"></param>
-        /// <returns></returns>
-        private bool CanSave(object property)
+        private void KeyUp(object parameter)
         {
-            return Owner.BitmapManager.ActiveDocument != null;
+            KeyEventArgs args = (KeyEventArgs)parameter;
+            if (_restoreToolOnKeyUp && Owner.ShortcutController.LastShortcut != null &&
+                Owner.ShortcutController.LastShortcut.ShortcutKey == args.Key)
+            {
+                _restoreToolOnKeyUp = false;
+                Owner.ToolsSubViewModel.SetActiveTool(Owner.ToolsSubViewModel.LastActionTool);
+                ShortcutController.BlockShortcutExecution = false;
+            }
         }
 
+        public void KeyDown(object parameter)
+        {
+            KeyEventArgs args = (KeyEventArgs)parameter;
+            if (args.IsRepeat && !_restoreToolOnKeyUp && Owner.ShortcutController.LastShortcut != null &&
+                Owner.ShortcutController.LastShortcut.Command == Owner.ToolsSubViewModel.SelectToolCommand)
+            {
+                _restoreToolOnKeyUp = true;
+                ShortcutController.BlockShortcutExecution = true;
+            }
+            Owner.ShortcutController.KeyPressed(args.Key, Keyboard.Modifiers);
+        }
     }
 }
